working copy - extract a workingCopy.ts file
